backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.26.0
sentence-transformers==2.2.2
scikit-learn==1.3.1
opencv-python==4.8.1.78
huggingface_hub==0.17.3